Added BaseAbility class and implemented passive ability system
[•] Added `BaseAbility` class to provide shared stats for `Ability` and `PassiveAbility`
[•] Implemented passive ability system with the first passive ability added
[•] Made minor adjustments across multiple classes due to the introduction of `BaseAbility`
[•] Optimized memory usage by updating data types in `Attributes.cs`

+++ Assets/_Scripts/Game/PlayerScripts/Attributes.cs
@@ -17,15 +17,15 @@ public static class Attributes
     public static float MaxHealth = 100;
     public static float PlayerDamageMult = 1f;
     public static float PlayerResistsMult = 1f;
-    public static int ProjectilesAdd = 0;
+    public static byte ProjectilesAdd = 0;
     public static float CastSpeedMult = 1f;
     public static float CooldownReductionMult = 1f;
-    public static int ActiveAbilitySlots = 5;
-    public static int AbilitiesPerLevel = 3;
-    public static int PassiveAbilitySlots = 5;
-    public static int ManuallyTriggeredAbilitySlots = 0;
+    public static byte ActiveAbilitySlots = 5;
+    public static byte AbilitiesPerLevel = 3;
+    public static byte PassiveAbilitySlots = 5;
+    public static byte ManuallyTriggeredAbilitySlots = 0;
     public static float PassiveAbilityEffectMult = 1f;
-    public static int PierceTargets = 0;
+    public static short PierceTargets = 0;
     public static event Action<AttributeId, float> OnAttributeChanged;
 
     private static float _magnetRadius = 0.5f;
